feat(lcd): 开机 BL 从 0 平滑渐亮到 80 (2s), g_ticks_ms 改 uint64 防溢出
- main.c: 移除开机固定亮度 50 + Codebot Logo 绘制, 改用 2s 内 0→80 渐亮
  (host 拉起后 standby 状态机接管, 不再启动即画 Logo)
- g_ticks_ms: uint32 → uint64 (49.7 天溢出风险; proto.c / hid_kbd.c 同步)

+++ include/pinout.h
@@ -89,7 +89,7 @@
 #include "ch32x035.h"
 
 /* ===== 全局工具 (在 main.c 中定义) ===== */
-extern volatile uint32_t g_ticks_ms;
+extern volatile uint64_t g_ticks_ms;
 void delay_ms(uint32_t ms);
 
 /* ===== GPIO 端口基地址 (CH32X035 头文件已定义 GPIOA/GPIOB/GPIOC) ===== */
